backend/requirements.txt
asgiref==3.8.1
astroid==3.3.5
attrs==24.2.0
autopep8==2.3.1
canvasapi==0.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
coverage==7.6.4
cryptography==43.0.3
defusedxml==0.8.0rc2
dill==0.3.9
Django==5.1.3
django-allauth==65.2.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.9.0
pylint==3.3.1
PySocks==1.7.1
pytest==8.3.3
pytest-cov==6.0.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.1
tomli==2.1.0
tomlkit==0.13.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
